changed score type -> isize

+++ src/slack/mod.rs
@@ -42,7 +42,10 @@ pub async fn notify_slack(
                 Language::Japanese,
             ));
 
-            let score = extracted_keywords.iter().map(|kwd| kwd.score).sum::<i8>();
+            let score = extracted_keywords
+                .iter()
+                .map(|kwd| kwd.score)
+                .sum::<isize>();
             if score <= 0 {
                 print!("Skipped: {} (score: {})\n", article.title, score);
                 excluded_articles.push((article.clone(), score as isize, extracted_keywords));
